Adjusting home_shots and away_shots
in FHM10, home_shots and away_shots are all shot attempts. Leading to confusion.

Adjust home_shots and away_shots in the api by having it be the sum of the shots from all 4 periods.
Add shot attempts to the shot quality table. As well as explaining in the modal on why