Feat/fhm 10 integration (#615)
* Init

* Updating boxscore to V3 and logic behind it

* Initial FHM10 boxscore

* Adding a modal to FAQ, have isFHM10 across all relevent modals

* Adding isHomeTeam and teamID to skaters and goalies

selfishly I want teamID in the skaters and goalies for my own use and experience using the API

* stepping up to v4

* Adjusting home_shots and away_shots

in FHM10, home_shots and away_shots are all shot attempts. Leading to confusion.

Adjust home_shots and away_shots in the api by having it be the sum of the shots from all 4 periods.
Add shot attempts to the shot quality table. As well as explaining in the modal on why

* Code Reivew changes